remove cycle
